# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Swimplan API
  description: REST API for the Swimplan swimming exercise generator.
  version: "0.1.0"

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /healthz:
    get:
      operationId: healthCheck
      summary: Health check
      tags: [system]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /api/v1/auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already taken

  /api/v1/auth/login:
    post:
      operationId: login
      summary: Log in and receive a JWT
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials

  /api/v1/exercises:
    get:
      operationId: listExercises
      summary: List all exercises
      tags: [exercises]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      summary: Create a new exercise
      tags: [exercises]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"

  /api/v1/exercises/{id}:
    get:
      operationId: getExercise
      summary: Get an exercise by ID
      tags: [exercises]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Exercise not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time

    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        description:
          type: string
        distance:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateExerciseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        distance:
          type: integer